requirements.txt
certifi             2024.8.30
charset-normalizer  3.4.0
colorama            0.4.6
contourpy           1.1.1
cycler              0.12.1
diffusers           0.31.0
filelock            3.16.1
fonttools           4.55.3
fsspec              2024.10.0
huggingface-hub     0.25.0
idna                3.10
imageio             2.35.1
importlib_metadata  8.5.0
importlib_resources 6.4.5
kiwisolver          1.4.7
lazy_loader         0.4
matplotlib          3.7.5
networkx            3.1
numpy               1.24.4
opencv-python       4.10.0.84
packaging           24.2
pillow              10.4.0
pip                 24.2
pycocotools         2.0.7
pyparsing           3.1.4
python-dateutil     2.9.0.post0
PyWavelets          1.4.1
PyYAML              6.0.2
regex               2024.11.6
requests            2.32.3
safetensors         0.4.5
scikit-image        0.21.0
scipy               1.10.1
segment_anything    1.0  
setuptools          75.1.0
six                 1.17.0
tifffile            2023.7.10
tokenizers          0.20.3
torch               1.11.0+cu113
torchaudio          0.11.0+cu113
torchvision         0.12.0+cu113
tqdm                4.67.1
transformers        4.46.3
typing_extensions   4.12.2
urllib3             2.2.3
wheel               0.44.0
zipp                3.20.2